+++ lint-staged.config.mjs
@@ -1,19 +1,23 @@
-import path from "node:path";
+import path from 'node:path'
 
-const buildEslintCommand = (filenames) =>
-  `next lint --fix --file ${filenames
-    .map((f) => path.relative(process.cwd(), f))
-    .join(" --file ")}`;
+const buildEslintCommand = (filenames) => {
+  const files = filenames
+    .map((f) => `"${path.relative(process.cwd(), f)}"`)
+    .join(' --file ')
+  return `cd apps/frontend && next lint --fix --file ${files}`
+}
 
 export default {
   // Run Next.js lint only on files in the Next.js app
-  "apps/frontend/**/*.{js,jsx,ts,tsx}": [
+  'apps/frontend/**/*.{js,jsx,ts,tsx}': [
     buildEslintCommand,
-    "prettier --write",
+    'prettier --write',
   ],
 
-  // Use general linting rules for other packages and files outside Next.js
-  "*.{js,jsx,ts,tsx}": ["eslint --fix", "prettier --write"],
-  "*.{md,mdx}": ["prettier --write"],
-  "*.{json,yaml,yml}": ["prettier --write"],
-};
+  // Use ESLint and Prettier only on apps (excluding frontend) and all packages
+  'apps/!(frontend)/**/*.{js,jsx,ts,tsx}': ['eslint --fix', 'prettier --write'],
+  'packages/**/*.{js,jsx,ts,tsx}': ['eslint --fix', 'prettier --write'],
+
+  '*.{md,mdx}': ['prettier --write'],
+  '*.{json,yaml,yml}': ['prettier --write'],
+}